# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RealmForge Control API",
    "description": "Governance kernel REST API — session, command, audit, snapshot, rollback, boards, and more.",
    "license": {
      "name": "UNLICENSED",
      "identifier": "UNLICENSED"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/snapshots": {
      "get": {
        "tags": [
          "snapshots"
        ],
        "summary": "List snapshots for the default project with pagination.",
        "operationId": "list_snapshots",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Snapshots listed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SnapshotResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "snapshots"
        ],
        "summary": "Create a new content-addressed snapshot manifest.",
        "operationId": "create_snapshot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSnapshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Snapshot created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/snapshots/compare": {
      "get": {
        "tags": [
          "snapshots"
        ],
        "summary": "Compare two snapshots and return the delta (added, removed, changed objects).",
        "operationId": "compare_snapshots",
        "parameters": [
          {
            "name": "from_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "snap_001"
          },
          {
            "name": "to_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "snap_002"
          }
        ],
        "responses": {
          "200": {
            "description": "Snapshot delta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotCompareResponse"
                }
              }
            }
          },
          "404": {
            "description": "One or both snapshots not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/snapshots/{id}": {
      "get": {
        "tags": [
          "snapshots"
        ],
        "summary": "Get a snapshot by ID, returning its validation report.",
        "operationId": "get_snapshot",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Snapshot ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Snapshot found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotValidateResponse"
                }
              }
            }
          },
          "404": {
            "description": "Snapshot not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/snapshots/{id}/validate": {
      "post": {
        "tags": [
          "snapshots"
        ],
        "summary": "Validate a snapshot's manifest hash integrity and object reference consistency.",
        "operationId": "validate_snapshot",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Snapshot ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Validation report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotValidateResponse"
                }
              }
            }
          },
          "404": {
            "description": "Snapshot not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/work-packets/generate": {
      "post": {
        "tags": [
          "work-packets"
        ],
        "summary": "Generate a scoped work packet for an AI agent from a work path node.",
        "operationId": "generate_work_packet",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateWorkPacketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Work packet generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkPacketGenerateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/work-packets/{id}/validate": {
      "post": {
        "tags": [
          "work-packets"
        ],
        "summary": "Validate a work packet's scope boundaries and cost budget.",
        "operationId": "validate_work_packet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Work packet ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Validation result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkPacketValidateResponse"
                }
              }
            }
          },
          "404": {
            "description": "Packet not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSnapshotRequest": {
        "type": "object",
        "required": [
          "tenant_id",
          "project_id",
          "reason"
        ],
        "properties": {
          "parent_snapshot_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "previous_manifest_hash": {
            "type": [
              "string",
              "null"
            ]
          },
          "project_id": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "tenant_id": {
            "type": "string"
          }
        }
      },
...
      "GenerateWorkPacketRequest": {
        "type": "object",
        "required": [
          "tenant_id",
          "project_id",
          "actor_id",
          "work_path_id",
          "node_id",
          "objective"
        ],
        "properties": {
          "actor_id": {
            "type": "string"
          },
          "cost_budget": {
            "type": [
              "object",
              "null"
            ]
          },
          "node_id": {
            "type": "string"
          },
          "objective": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          },
          "tenant_id": {
            "type": "string"
          },
          "work_path_id": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 7807 Problem Details response, extended with a machine-readable error code\nand trace ID per the RealmForge API error standard.",
        "required": [
          "code",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable SCREAMING_SNAKE_CASE error code. Stable across versions."
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "trace_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Correlation ID matching the X-Request-Id response header."
          }
        }
      },
...
      "SnapshotCompareResponse": {
        "type": "object",
        "required": [
          "delta"
        ],
        "properties": {
          "delta": {
            "type": "object"
          }
        }
      },
      "SnapshotResponse": {
        "type": "object",
        "required": [
          "snapshot_id",
          "status",
          "reason",
          "object_count"
        ],
        "properties": {
          "object_count": {
            "type": "integer",
            "minimum": 0
          },
          "reason": {
            "type": "string"
          },
          "snapshot_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "SnapshotValidateResponse": {
        "type": "object",
        "required": [
          "report"
        ],
        "properties": {
          "report": {
            "type": "object"
          }
        }
      },
...
      "WorkPacketGenerateResponse": {
        "type": "object",
        "required": [
          "packet_id",
          "status"
        ],
        "properties": {
          "packet_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "WorkPacketValidateResponse": {
        "type": "object",
        "required": [
          "result"
        ],
        "properties": {
          "result": {
            "type": "object"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "session_token"
      }
    }
  },
...
}